Repalce take_data_file.
