+++ pylabnet/scripts/data_center/take_data.py
@@ -14,6 +14,7 @@ from pylabnet.utils.logging.logger import LogHandler
 from pylabnet.gui.pyqt.external_gui import Window
 from pylabnet.utils.helper_methods import load_config, generic_save, unpack_launcher, save_metadata, load_script_config, find_client, get_ip
 from pylabnet.scripts.data_center import datasets
+from PyQt5.QtWidgets import QLabel, QLineEdit, QPushButton
 
 
 REFRESH_RATE = 150   # refresh rate in ms, try increasing if GUI lags
@@ -28,7 +29,7 @@ class DataTaker:
 
         # Instantiate GUI window
         self.gui = Window(
-            gui_template='data_taker',
+            gui_template='data_taker_input',
             host=get_ip(),
             log=self.log,
         )
@@ -126,6 +127,72 @@ class DataTaker:
             self.cur_path = self.gui.exp.model().filePath(self.gui.exp.currentIndex())
             self.exp_name = os.path.split(os.path.basename(self.cur_path))[1][:-3]
 
+        self.load_init_dict(index)
+
+
+    def load_init_dict(self, index):
+        """ Displays the currently clicked experiment's init dict
+
+        :param index: index of (QTreeView) entry to display
+        """
+
+        filepath = self.gui.exp.model().filePath(index)
+        if not os.path.isdir(filepath):
+           
+            spec = importlib.util.spec_from_file_location("init_dict", filepath)
+            mod = importlib.util.module_from_spec(spec)
+            spec.loader.exec_module(mod)
+            self.init_dict = mod.INIT_DICT
+
+        self.update_init_form(self.init_dict)
+
+    def create_gridlayout(self):
+
+        input_init_scrollarea = self.gui.init_dict_scrollarea
+
+        self.gui.scrollAreaWidgetContents = QtWidgets.QWidget()
+        self.gui.input_dict_gridlayout = QtWidgets.QGridLayout(self.gui.scrollAreaWidgetContents)
+        input_init_scrollarea.setWidget(self.gui.scrollAreaWidgetContents)
+
+    def update_init_form(self, newform):
+       
+        self.create_gridlayout()
+
+        # Build input dict modification form
+        input_grid_layout = self.gui.input_dict_gridlayout
+        btn1 = QPushButton("Push parameters")
+        
+        for i,  (labelname, default_value) in enumerate(newform.items()):
+            
+                label = QLabel(labelname)
+                value_entry =  QLineEdit(default_value)
+                input_grid_layout.addWidget(label, i, 0)
+                input_grid_layout.addWidget(value_entry, i, 1)
+
+
+    def load_input_dict(self):
+        """ Load input dict from Gridlayout in GUI"""
+
+         # load input dict
+        input_grid_layout = self.gui.input_dict_gridlayout
+        input_scrollarea = self.gui.scrollAreaWidgetContents
+
+        num_cols = input_grid_layout.columnCount()
+        num_rows = input_grid_layout.rowCount()
+        init_widgets = input_scrollarea.children()[1:] # omitting first one since it's the gridlayout 
+
+        assert len(init_widgets) == num_cols * num_rows
+
+        reconstructed_init_dict = {}    
+
+        for i in range(num_rows):
+            reconstructed_init_dict[init_widgets[i*2].text()] = float(init_widgets[i*2+ 1].text())
+
+        self.log.info(f'Input Dict {reconstructed_init_dict} configured')
+        # Commit input dict to parent dataset
+        self.dataset.set_input_dict(reconstructed_init_dict)
+
+
     def configure(self):
         """ Configures the currently selected experiment + dataset """
 
@@ -176,12 +243,16 @@ class DataTaker:
             self.log.error(error_msg)
             return
 
+        # Load input dict from GUI so it is accessible for config function.
+        self.load_input_dict()
+
         # Run any pre-experiment configuration
         try:
             self.module.configure(dataset=self.dataset, **self.clients)
         except AttributeError:
             pass
         self.experiment = self.module.experiment
+    
 
         self.log.info(f'Experiment {self.exp_name} configured')
         self.gui.exp_preview.setStyleSheet('font: 10pt "Consolas"; '
